Fixed portability issues

+++ src/pseudo.h
@@ -7,11 +7,13 @@
 #ifndef __PSEUDO_H__
 #define __PSEUDO_H__
 
+#include <stdint.h>
+
 //PseudoHeader struct used to calculate UDP checksum.
 typedef struct PseudoHeader
 {
-    unsigned long int source_ip;
-    unsigned long int dest_ip;
+    uint32_t source_ip;
+    uint32_t dest_ip;
     unsigned char reserved;
     unsigned char protocol;
     unsigned short int udp_length;
